feat: Update Signup action to hash user password before creating user
This commit updates the Signup action in the actions.js file to hash the user's password before creating a new user. This is done to improve security and protect user passwords. If an error occurs during the creation of the user and the error code is "SQLITE_CONSTRAINT_UNIQUE", it returns an error message indicating that the email is already in use. Otherwise, it throws the error.